prepare for --fix thing
